+++ packages/webview_flutter/webview_flutter_wkwebview/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 3.17.0
 
+* Adds a change listener for the `canGoBack` property. See
+  `WebKitWebViewController.setOnCanGoBackChange`.
 * Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
 * Updates internal API wrapper to use ProxyApis.